rename logging to structured_logging
